Reference

Reference

Steady state solver

Plans

Stacked time solver

Various

Index

Steady state solver

StateSpaceEcon.SteadyStateSolver — Module

SteadyStateSolver

A module that is part of StateSpaceEcon package. Contains methods for finding a steady state of a model.

StateSpaceEcon.SteadyStateSolver.check_sstate — Method

check_sstate(model; <options>)

Run a diagnostic test to determine if the steady state solution stored within the given model object is indeed a solution of the steady state system of equations.

Return the number of steady state equations that are violated by the current steady state solution. If verbose=true, also display diagnostic information in the form of listing all bad equations and their residuals.

Options

Standard options (default values from model.options)

verbose

tol - an equation is considered satisfied if its residual, in absolute value, is smaller than this number.

StateSpaceEcon.SteadyStateSolver.clear_sstate! — Method

clear_sstate!(model; lvl=0.1, slp=0.0, <options>)

Set the steady state values to the provided defaults and presolve.

Arguments

model - the model instance

lvl, slp - the initial guess for the level and the slope. Each could be a number or a vector of length equal to the number of variable in the mode.

Options

Standard options (default values are taken from model.options)

verbose

StateSpaceEcon.SteadyStateSolver.diagnose_sstate — Method

diagnose_sstate([point,] model)

Run diagnostics on the steady state of the given model. If point is not given, then we check the steady state solution stored inside the given model.

Return a tuple of "bad" equations and "bad" variables.

The set of "bad" equations is one that is inconsistent, i.e. there is no solution. This might happen if the system is overdetermined.

The set of "bad" variables contains variables that cannot be solved uniquely. This might happen if the system is underdetermined. In this case, try adding steady state constraints until you get a unique solution. See @steadystate in ModelBaseEcon.

Internal function

The output from this function may be difficult to read.<br> Call check_sstate instead.

StateSpaceEcon.SteadyStateSolver.initial_sstate! — Method

initial_sstate!(model, init; <options>)

Set the steady state values from the given vector and presolve.

Call this function to specify initial guesses for the iterative steady state solver. If the value of a steady state variable is known, it is better to use @steadystate to add that as a steady state constraint.

Arguments

model - the model.

init - a vector of length equal to twice the number of variables in the model. The level and slope values are staggered, i.e., the level and slope of variable j are in init[2j-1] and init[2j].

Options

Standard options (default values are taken from model.options)

verbose

StateSpaceEcon.SteadyStateSolver.sssolve! — Method

sssolve!(model; <options>)

Solve the steady state problem for the given model.

Options

Standard options (default values are taken from model.options)

verbose

tol, maxiter - control the stopping criteria of the solver

Specific options

presolve::Bool - whether or not to use a presolve pass. Default is true.

method::Symbol - choose the solution algorithm. Valid options are :nr for Newton-Raphson, :lm for Levenberg-Marquardt, and :auto. The :auto method starts with the LM algorithm and automatically switches to NR when it starts to converge. Default is :nr.

Plans

StateSpaceEcon.Plans — Module

Plans

Module part of StateSpaceEcon. This module implements the Plan data structure, which is used in simulations. The plan object contains information about the range of the simulation and which variables and shocks are exogenous or endogenous at each period of the range.

Constructors

Plan(model, range)

Modify the plan

exogenize!, endogenize! - make variables exogenous or endogenous

exog_endo!, endo_exog! - swap exogenous and endogenous variables

autoexogenize! - exogenize and endogenize variables according to the list in the model

Prepare data for simulation

zeroarray, zerodict, zerodata - prepare a matrix or a dictionary or a SimData of data for the simulation containing zeros.

steadystatearray, steadystatedict, steadystatedata - prepare a matrix or a dictionary or a SimData of data for the simulation containing the steady state.

StateSpaceEcon.Plans.Plan — Type

Plan{T <: MIT}

A data structure representing the simulation plan. It holds information about the time range of the simulation and which variables/shocks are exogenous at each period.

StateSpaceEcon.Plans.Plan — Method

Plan(model, range)

Create a default simulation plan for the given model over the given range. The range of the plan is augmented to include periods before and after the given range, over which initial and final conditions will be applied.

Instead of a range, one could also pass in a single moment in time (MIT) instance, in which case it is interpreted as a range of length 1.

StateSpaceEcon.Plans.autoexogenize! — Method

autoexogenize!(plan, model, date)

Modify the given plan according to the "autoexogenize" protocol defined in the given model. All variables in the autoexogenization list become endogenous and their corresponding shocks become exogenous over the given date or range. date can be a moment in time (same frequency as the given plan), a range, an iterable, or a container.

StateSpaceEcon.Plans.endo_exog! — Method

endo_exog!(plan, endo_vars, exog_vars, date)

Modify the given plan so that the given variables listed in exog_vars will be exogenous and the variables listed in endo_vars will be endogenous on the given dates. exog_vars and endo_vars can each be a Symbol or a String or a Vector of such. date can be a moment in time (same type as the plan), or a range or an iterable or a container.

StateSpaceEcon.Plans.endogenize! — Method

endogenize!(plan, vars, date)

Modify the given plan so that the given variables will be endogenous on the given dates. vars can be a Symbol or a String or a Vector of such. date can be a moment in time (same type as the plan), or a range or an iterable or a container.

StateSpaceEcon.Plans.exog_endo! — Method

exog_endo!(plan, exog_vars, endo_vars, date)

Modify the given plan so that the given variables listed in exog_vars will be exogenous and the variables listed in endo_vars will be endogenous on the given dates. exog_vars and endo_vars can each be a Symbol or a String or a Vector of such. date can be a moment in time (same type as the plan), or a range or an iterable or a container.

StateSpaceEcon.Plans.exogenize! — Method

exogenize!(plan, vars, date)

Modify the given plan so that the given variables will be exogenous on the given dates. vars can be a Symbol or a String or a Vector of such. date can be a moment in time (same type as the plan), or a range or an iterable or a container.

StateSpaceEcon.Plans.plansum — Method

plansum(model, plan)

Return the total number of exogenous variables in the simulation plan. Periods over which initial and final conditions are imposed are not counted in this sum.

StateSpaceEcon.Plans.setexog! — Method

setexog!(plan, t, vinds)

Modify the plan at time t such that vinds are exogenous and the rest are endogenous.

StateSpaceEcon.Plans.steadystatearray — Method

steadystatearray(model, plan)
steadystatearray(model, range)

Create a matrix of the proper dimensions for a simulation with the given model with the given plan or over the given range. The matrix is initialized with the steady state level of each variable. If a range is given rather than a plan, it is augmented with periods before and after the given range in order to accommodate initial and final conditions.

See also: zeroarray, zerodict, steadystatearray, steadystatedict

StateSpaceEcon.Plans.steadystatedata — Method

steadystatedata(model, plan)
steadystatedata(model, range)

Create a SimData containing a TSeries of the appropriate range for each variable in the model for a simulation with the given plan or over the given range. The matrix is initialized with the steady state level of each variable. If a range is given rather than a plan, it is augmented with periods before and after the given range in order to accommodate initial and final conditions.

See also: zeroarray, zerodict, steadystatearray, steadystatedict

StateSpaceEcon.Plans.steadystatedict — Method

steadystatearray(model, plan)
steadystatearray(model, range)

Create a dictionary containing a TSeries of the appropriate range for each variable in the model for a simulation with the given plan or over the given range. The matrix is initialized with the steady state level of each variable. If a range is given rather than a plan, it is augmented with periods before and after the given range in order to accommodate initial and final conditions.

See also: zeroarray, zerodict, steadystatearray, steadystatedict

StateSpaceEcon.Plans.zeroarray — Method

zeroarray(model, plan)
zeroarray(model, range)

Create a matrix of the proper dimension for a simulation with the given model with the given plan or over the given range. If a range is given, the data is prepared for the default plan. This means that appropriate number of periods are added before and after the range to account for initial and final conditions.

See also: zeroarray, zerodict, steadystatearray, steadystatedict

StateSpaceEcon.Plans.zerodata — Method

zerodata(model, plan)
zerodata(model, range)

Create a NamedTuple containing a TSeries of the appropriate range for each variable in the model for a simulation with the given plan or over the given range. If a range is given rather than a plan, the data is prepared for the default plan over that range. This means that appropriate number of periods are added before and after the range to account for initial and final conditions.

See also: zeroarray, zerodict, steadystatearray, steadystatedict

StateSpaceEcon.Plans.zerodict — Method

zerodict(model, plan)
zerodict(model, range)

Create a dictionary containing a TSeries of the appropriate range for each variable in the model for a simulation with the given plan or over the given range. If a range is given rather than a plan, the data is prepared for the default plan over that range. This means that appropriate number of periods are added before and after the range to account for initial and final conditions.

See also: zeroarray, zerodict, steadystatearray, steadystatedict

Stacked time solver

StateSpaceEcon.StackedTimeSolver — Module

StackedTimeSolver

A module that is part of StateSpaceEcon package. Contains methods for solving the dynamic system of equations for the model and running simulations.

StateSpaceEcon.StackedTimeSolver.FCType — Type

FCType

An enumeration type representing the possible types of final conditions.

fcgiven - the final conditions are provided by the user.

fclevel - the final conditions are prescribed from the levels of the steady state solution.

fcslope - the final conditions are such that the slope of the dynamic solution matches the slope of the steady state solution.

StateSpaceEcon.StackedTimeSolver.array2data — Method

array2data(data, vars, start_date)

Convert the simulation data array to a named tuple.

StateSpaceEcon.StackedTimeSolver.array2dict — Method

array2dict(data, vars, start_date)

Convert the simulation data array to a dictionary.

StateSpaceEcon.StackedTimeSolver.dict2array — Method

dict2array(d, vars; range)

Convert a dictionary of TSeries to a 2d array of simulation data for the given range. The range argument is optional and defaults to nothing.

StateSpaceEcon.StackedTimeSolver.dictoverlay — Method

dictoverlay(d1, d2)

Merge two dictionaries. Common key where the values are TSeries of the same frequency are overlayed. Otherwise, a common key takes the value of the last Dict containing it.

See also: seriesoverlay

StateSpaceEcon.StackedTimeSolver.seriesoverlay — Method

seriesoverlay(ts1, ts2)

Return a new TSeries over the full range of both arguments. The overlapping part contains values from the last argument.

See also: dictoverlay

StateSpaceEcon.StackedTimeSolver.simulate — Method

simulate(model, plan, data; <options>)

Run a simulation for the given model, simulation plan and exogenous data.

Arguments

model - the Model instance to simulate.

plan - the Plan for the simulation.

data - a 2D Array containing the exogenous data. This includes the initial and final conditions.

Options as keyword arguments

fctype::FCType - set the desired final condition type for the simulation. The default value is fcgiven. Other possible values include fclevel and fcslope.

initial_guess::AbstractMatrix{Float64} - a 2D Array containing the initial guess for the solution. This is used to start the Newton-Raphson algorithm. The default value is an empty array (zeros(0,0)), in which case we use the exogenous data for the initial condition. You can use the steady state solution using steadystatearray.

linearize::Bool - set to true to instruct the solver to use the liearized model. If the model is already linearized, this option has the effect that the model gets linearized about the current steady stat and with the value of deviation given here. Otherwise the model is linearized about the steady state. After the simulation is computed, the model is restored to its original state. Default value is false.

deviation::Bool - set to true if the data is in deviations from the steady state. This is only relevant if the linearize option is set to true. Default value is false.

anticipate::Bool - set to false to instruct the solver that all shocks are unanticilated by the agents. Default value is true.

verbose::Bool - control whether or not to print progress information. Default value is taken from model.options.

tol::Float64 - set the desired accuracy. Default value is taken from model.options.

maxiter::Int - algorithm fails if the desired accuracy is not reached within this maximum number of iterations. Default value is taken from model.options.

See also:

Examples

Various

StateSpaceEcon.StateSpaceEcon — Module

StateSpaceEcon

A package for Macroeconomic modelling.

StateSpaceEcon.SimData — Type

SimData

Data structure containing the time series data for a simulation.

It is a collection of TSeries of the same frequency and containing data for the same range. When used for simulation, the range must include the initial conditions, the simulation range and the final conditions, although it could extend beyond that. It must contain time series for all variables and shocks in the model, although it might contain other time series.

StateSpaceEcon.printmatrix — Method

printmatrix(mat [, Val(F), colnames])

Display a matrix in full while controlling the formatting of each value and optionally showing the column names.

Val(F) - display each number in the given format F. The format is in the form of a decimal point number where the whole part indicates the total width and the fractional part is the number of digits printed after the decimal point. Default is Val(12.7)

colnames - a list of names to display in the first row. The names are displayed as given, possibly with padding to match the width given in the Val argument. If any names are longer than that, they will not be truncated and so the display will not be aligned properly. Sorry about that!

Index

StateSpaceEcon.Plans

StateSpaceEcon.StackedTimeSolver

StateSpaceEcon.StateSpaceEcon

StateSpaceEcon.SteadyStateSolver

StateSpaceEcon.Plans.Plan

StateSpaceEcon.Plans.Plan

StateSpaceEcon.SimData

StateSpaceEcon.StackedTimeSolver.FCType

StateSpaceEcon.Plans.autoexogenize!

StateSpaceEcon.Plans.endo_exog!

StateSpaceEcon.Plans.endogenize!

StateSpaceEcon.Plans.exog_endo!

StateSpaceEcon.Plans.exogenize!

StateSpaceEcon.Plans.plansum

StateSpaceEcon.Plans.setexog!

StateSpaceEcon.Plans.steadystatearray

StateSpaceEcon.Plans.steadystatedata

StateSpaceEcon.Plans.steadystatedict

StateSpaceEcon.Plans.zeroarray

StateSpaceEcon.Plans.zerodata

StateSpaceEcon.Plans.zerodict

StateSpaceEcon.StackedTimeSolver.array2data

StateSpaceEcon.StackedTimeSolver.array2dict

StateSpaceEcon.StackedTimeSolver.dict2array

StateSpaceEcon.StackedTimeSolver.dictoverlay

StateSpaceEcon.StackedTimeSolver.seriesoverlay

StateSpaceEcon.StackedTimeSolver.simulate

StateSpaceEcon.SteadyStateSolver.check_sstate

StateSpaceEcon.SteadyStateSolver.clear_sstate!

StateSpaceEcon.SteadyStateSolver.diagnose_sstate

StateSpaceEcon.SteadyStateSolver.initial_sstate!

StateSpaceEcon.SteadyStateSolver.sssolve!

StateSpaceEcon.printmatrix